refactor : PostRepositoryImpl 리펙토링 #113
