+++ src/main/java/com/backend/naildp/controller/HomeController.java
@@ -12,7 +12,7 @@ import org.springframework.web.bind.annotation.RestController;
 
 import com.backend.naildp.dto.home.PostSummaryResponse;
 import com.backend.naildp.exception.ApiResponse;
-import com.backend.naildp.service.PostInfoService;
+import com.backend.naildp.service.post.PostInfoService;
 
 import lombok.RequiredArgsConstructor;
