+++ src/main/java/com/backend/naildp/repository/PostRepositoryImpl.java
@@ -25,6 +25,7 @@ import com.querydsl.core.types.dsl.StringExpressions;
 import com.querydsl.jpa.JPAExpressions;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 
+import io.jsonwebtoken.lang.Strings;
 import jakarta.persistence.EntityManager;
 
 public class PostRepositoryImpl implements PostSearchRepository {
@@ -41,9 +42,9 @@ public class PostRepositoryImpl implements PostSearchRepository {
 			.select(post)
 			.from(post)
 			.where(
-				usernamePermitted(username)
+				isAllowedToViewPosts(username)
 					.and(containsInPost(keywords))
-					.and(lessThanCursor(cursorId))
+					.and(lessThanCustomCursor(cursorId))
 			)
 			.orderBy(post.postLikes.size().desc(), post.createdDate.desc())
 			.limit(pageable.getPageSize() + 1)
@@ -78,7 +79,10 @@ public class PostRepositoryImpl implements PostSearchRepository {
 		return false;
 	}
 
-	private BooleanExpression usernamePermitted(String usernameCond) {
+	private BooleanExpression isAllowedToViewPosts(String usernameCond) {
+		if (!Strings.hasText(usernameCond)) {
+			return null;
+		}
 
 		return post.boundary.eq(Boundary.ALL)
 			.or(post.boundary.eq(Boundary.FOLLOW).and(
@@ -121,7 +125,7 @@ public class PostRepositoryImpl implements PostSearchRepository {
 			.contains(keyword);
 	}
 
-	private BooleanExpression lessThanCursor(Long cursorId) {
+	private BooleanExpression lessThanCustomCursor(Long cursorId) {
 		if (cursorId == null) {
 			return null;
 		}
@@ -129,21 +133,20 @@ public class PostRepositoryImpl implements PostSearchRepository {
 		DateTimeTemplate<String> stringDateTimeTemplate = Expressions.dateTimeTemplate(String.class,
 			"DATE_FORMAT({0}, '%y%m%d%H%i%s')", post.createdDate);
 
-		String cursor = generateCursor(cursorId, stringDateTimeTemplate);
+		String cursor = generateCustomCursor(cursorId, stringDateTimeTemplate);
 
-		return generateDateTimeFormat(stringDateTimeTemplate)
-			.lt(cursor);
+		return generateCustomExpression(stringDateTimeTemplate).lt(cursor);
 	}
 
-	private String generateCursor(Long cursorId, DateTimeTemplate<String> stringDateTimeTemplate) {
+	private String generateCustomCursor(Long cursorId, DateTimeTemplate<String> stringDateTimeTemplate) {
 		return queryFactory
-			.select(generateDateTimeFormat(stringDateTimeTemplate))
+			.select(generateCustomExpression(stringDateTimeTemplate))
 			.from(post)
 			.where(post.id.eq(cursorId))
 			.fetchOne();
 	}
 
-	private StringExpression generateDateTimeFormat(DateTimeTemplate<String> stringDateTimeTemplate) {
+	private StringExpression generateCustomExpression(DateTimeTemplate<String> stringDateTimeTemplate) {
 		return StringExpressions.lpad(post.postLikes.size().stringValue(), 6, '0')
 			.concat(StringExpressions.lpad(stringDateTimeTemplate.stringValue(), 12, '0')
 				.concat(StringExpressions.lpad(post.id.stringValue(), 8, '0')));
