+++ src/test/java/com/backend/naildp/service/PostInfoServiceTest.java
@@ -34,6 +34,7 @@ import com.backend.naildp.repository.PostLikeRepository;
 import com.backend.naildp.repository.PostRepository;
 import com.backend.naildp.repository.SocialLoginRepository;
 import com.backend.naildp.repository.UserRepository;
+import com.backend.naildp.service.post.PostInfoService;
 
 import jakarta.persistence.EntityManager;
